Rearrange birelay traits
