+++ crates/relayer-all-in-one/src/one_for_all/impls/builder/mod.rs
@@ -2,6 +2,7 @@ pub mod batch;
 pub mod birelay;
 pub mod cache;
 pub mod chain;
+pub mod component;
 pub mod error;
 pub mod logger;
 pub mod relay;
