+++ crates/relayer-components-extra/src/builder/components/relay/batch.rs
@@ -1,5 +1,5 @@
 use async_trait::async_trait;
-use ibc_relayer_components::builder::traits::birelay::HasBiRelayType;
+use ibc_relayer_components::builder::traits::birelay::types::HasBiRelayType;
 use ibc_relayer_components::builder::traits::relay::from_chains::RelayFromChainsBuilder;
 use ibc_relayer_components::builder::traits::target::chain::ChainBuildTarget;
 use ibc_relayer_components::builder::traits::target::relay::RelayBuildTarget;
