+++ crates/relayer-components/src/builder/impls/birelay.rs
@@ -1,6 +1,7 @@
 use async_trait::async_trait;
 
-use crate::builder::traits::birelay::{BiRelayBuilder, CanBuildBiRelayFromRelays};
+use crate::builder::traits::birelay::build::BiRelayBuilder;
+use crate::builder::traits::birelay::from_relays::CanBuildBiRelayFromRelays;
 use crate::builder::traits::relay::build::CanBuildRelay;
 use crate::builder::traits::target::relay::{RelayAToBTarget, RelayBToATarget};
 use crate::builder::types::aliases::{ChainIdA, ChainIdB, ClientIdA, ClientIdB};
