+++ crates/relayer-components/src/builder/impls/bootstrap/birelay.rs
@@ -1,7 +1,8 @@
 use async_trait::async_trait;
 
 use crate::builder::impls::bootstrap::relay::CanBootstrapRelay;
-use crate::builder::traits::birelay::{CanBuildBiRelay, HasBiRelayType};
+use crate::builder::traits::birelay::build::CanBuildBiRelay;
+use crate::builder::traits::birelay::types::HasBiRelayType;
 use crate::builder::traits::target::relay::RelayAToBTarget;
 use crate::builder::types::aliases::{ChainA, ChainB, ChainIdA, ChainIdB};
 use crate::chain::traits::client::create::HasCreateClientOptions;
