+++ crates/relayer-components/src/builder/traits/birelay/from_relays.rs
@@ -1,0 +1,15 @@
+use async_trait::async_trait;
+
+use crate::builder::traits::birelay::types::HasBiRelayType;
+use crate::builder::types::aliases::{RelayAToB, RelayBToA};
+use crate::core::traits::error::HasErrorType;
+use crate::std_prelude::*;
+
+#[async_trait]
+pub trait CanBuildBiRelayFromRelays: HasBiRelayType + HasErrorType {
+    async fn build_birelay_from_relays(
+        &self,
+        relay_a_to_b: RelayAToB<Self>,
+        relay_b_to_a: RelayBToA<Self>,
+    ) -> Result<Self::BiRelay, Self::Error>;
+}
