+++ crates/relayer-components/src/builder/traits/cache.rs
@@ -1,4 +1,4 @@
-use crate::builder::traits::birelay::HasBiRelayType;
+use crate::builder::traits::birelay::types::HasBiRelayType;
 use crate::builder::traits::target::chain::ChainBuildTarget;
 use crate::builder::traits::target::relay::RelayBuildTarget;
 use crate::builder::types::aliases::{TargetChainCache, TargetRelayCache};
